added a bunch of tests, refactored around tests - also in mid-refactor around the sample service workarounds - the disabled a couple of tests which need to be redone after the refactoring...
