+++ react-app/src/components/Comparator/style.less
