+++ react-app/src/components/DataLinks/view.test-data.json
@@ -1,0 +1,62 @@
+[
+    {
+        "linkid": "d8c05c70-4212-4e39-9b73-b385133351a2",
+        "upa": "51169/6/8",
+        "dataid": null,
+        "id": "8cc90218-51ae-4814-9c51-80ae990661e5",
+        "version": 1,
+        "node": "IEAWH006A.3",
+        "createdby": "eapearson",
+        "created": 1591793927329,
+        "expiredby": null,
+        "expired": null,
+        "key": "51169/6/8",
+        "objectName": "rhodobacter_CACIA14H1.reference",
+        "objectType": "KBaseAssembly.ReferenceAssembly-1.1"
+    },
+    {
+        "linkid": "219cfdb6-ce90-4941-bca9-f17a7675e784",
+        "upa": "51169/5/1",
+        "dataid": null,
+        "id": "8cc90218-51ae-4814-9c51-80ae990661e5",
+        "version": 1,
+        "node": "IEAWH006A.3",
+        "createdby": "eapearson",
+        "created": 1591794052522,
+        "expiredby": null,
+        "expired": null,
+        "key": "51169/5/1",
+        "objectName": "Rhodobacter_CACIA_14H1",
+        "objectType": "KBaseGenomes.Genome-7.0"
+    },
+    {
+        "linkid": "d3e042cf-31cd-40a7-a80b-1a9641986cd3",
+        "upa": "51169/4/8",
+        "dataid": null,
+        "id": "8cc90218-51ae-4814-9c51-80ae990661e5",
+        "version": 1,
+        "node": "IEAWH006A.3",
+        "createdby": "eapearson",
+        "created": 1591794068884,
+        "expiredby": null,
+        "expired": null,
+        "key": "51169/4/8",
+        "objectName": "Rsp-minimal",
+        "objectType": "KBaseBiochem.Media-1.0"
+    },
+    {
+        "linkid": "84b21635-88b0-4010-8c0d-ef581fd1e2f5",
+        "upa": "51169/2/1",
+        "dataid": "some.path.into.object?",
+        "id": "8cc90218-51ae-4814-9c51-80ae990661e5",
+        "version": 1,
+        "node": "IEAWH006A.3",
+        "createdby": "eapearson",
+        "created": 1591794085982,
+        "expiredby": "eapearson",
+        "expired": 1591794085982,
+        "key": "51169/2/1",
+        "objectName": "floc_mini_metadata_sample_set",
+        "objectType": "KBaseSets.SampleSet-2.0"
+    }
+]
