--- react-app/src/components/History/Comparator.css
@@ -1,34 +1,0 @@
-.Comparator {
-    flex: 1 1 0px;
-    display: flex;
-    flex-direction: column;
-    min-height: 0;
-}
-
-.Comparator .-diff {
-    background-color: aqua;
-}
-
-.Comparator .-viewSelector {
-    /* background-color: silver; */
-    margin-bottom: 10px;
-    text-align: center;
-}
-
-.IconStack {
-    position: relative;
-    color: red;
-}
-
-.IconStack .-icon {
-    position: absolute;
-    left: 0;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-}
