+++ react-app/src/components/History/view.test.tsxx
@@ -1,0 +1,58 @@
+import React from 'react';
+import { render, waitFor, findByRole, within } from '@testing-library/react';
+
+import HistoryTool, { HistoryToolProps } from './view';
+import testData from './view.test.data.json';
+
+const TIMEOUT = 10000;
+
+test('renders empty history', async () => {
+    const props: HistoryToolProps = {
+        history: [],
+        fieldDefinitions: {}
+    };
+
+    const { getByText, findByText } = render(<HistoryTool {...props} />);
+
+    const historyTitle = await findByText('History');
+    expect(historyTitle).toBeInTheDocument();
+
+    const diffTitle = await findByText('Diff');
+    expect(diffTitle).toBeInTheDocument();
+
+    // await waitFor(() => {
+    //     const historyTitle = getByText('History');
+    //     expect(historyTitle).toBeInTheDocument();
+    //     const diffTitle = getByText('Diff');
+    //     expect(diffTitle).toBeInTheDocument();
+    // }, {
+    //     timeout: TIMEOUT
+    // });
+});
+
+test('renders history', async () => {
+    const props: HistoryToolProps = testData as HistoryToolProps;
+
+    const { getByText, findByText, findByRole } = render(<HistoryTool {...props} />);
+
+    const titleRow = await findByRole('');
+
+    const withinTitleRow = within(titleRow);
+
+
+    const historyTitle = await withinTitleRow.findByText('History');
+    expect(historyTitle).toBeInTheDocument();
+
+    const diffTitle = await withinTitleRow.findByText('Diff');
+    expect(diffTitle).toBeInTheDocument();
+
+    // await waitFor(() => {
+    //     const historyTitle = getByText('History');
+    //     expect(historyTitle).toBeInTheDocument();
+    //     const diffTitle = getByText('Diff');
+    //     expect(diffTitle).toBeInTheDocument();
+    // }, {
+    //     timeout: TIMEOUT
+    // });
+});
+
