+++ react-app/src/components/Main/data.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { AsyncProcess, AsyncProcessStatus } from '../../redux/store/processing';
 import SampleServiceClient, {
-    SampleId, EpochTimeMS, SampleVersion, Username, FieldDefinition
+    SampleId, SampleVersion, Username
 } from '../../lib/comm/dynamicServices/SampleServiceClient';
 import { AppError } from '@kbase/ui-components';
 import Component from './view';
@@ -9,70 +9,7 @@ import { LoadingOutlined } from '@ant-design/icons';
 import { Alert } from 'antd';
 import { UPSTREAM_TIMEOUT } from '../../constants';
 import UserProfileClient from '../../lib/comm/coreServices/UserProfileClient';
-
-export interface MetadataField {
-    value: string | number | boolean;
-    units?: string;
-    label: string;
-    description?: string;
-    isControlled: boolean;
-    definition: FieldDefinition;
-}
-
-export interface UserMetadataField {
-    value: string;
-}
-
-export interface Metadata {
-    [key: string]: MetadataField;
-}
-
-export interface UserMetadata {
-    [key: string]: UserMetadataField;
-}
-
-export interface User {
-    username: Username;
-    realname: string;
-    gravatarHash: string;
-    avatarOption?: string;
-    gravatarDefault?: string;
-}
-
-export interface ACL {
-    admin: Array<User>;
-    write: Array<User>;
-    read: Array<User>;
-}
-
-
-export type SampleType = 'BioReplicate' | 'TechReplicate' | 'SubSample';
-
-export interface Sample {
-    id: SampleId;
-    name: string;
-    created: {
-        at: EpochTimeMS;
-        by: User;
-    };
-    currentVersion: {
-        at: EpochTimeMS;
-        by: User;
-        version: number;
-    };
-    latestVersion: {
-        at: EpochTimeMS;
-        by: User;
-        version: number;
-    };
-    source: string;
-    sourceId: string;
-    sourceParentId: string | null;
-    type: SampleType;
-    metadata: Metadata;
-    userMetadata: UserMetadata;
-}
-
+import { Sample, Metadata, User } from './types';
 
 export interface DataProps {
     serviceWizardURL: string;
