+++ react-app/src/components/Main/style.less
@@ -21,17 +21,17 @@
     cursor: pointer;
 }
 
-.InfoTable.UserMetadata>div>div:nth-child(1) {
+.Main .InfoTable.UserMetadata>div>div:nth-child(1) {
     flex: 0 0 20em;
 }
 
-.InfoTable.UserMetadata>div>div:nth-child(2) {
+.Main .InfoTable.UserMetadata>div>div:nth-child(2) {
     flex: 0 0 20em;
     color: gray;
     font-style: italic;
 }
 
-.InfoTable.ControlledMetadata>div>div:nth-child(1) {
+.Main .InfoTable.ControlledMetadata>div>div:nth-child(1) {
     flex: 0 0 20em;
 }
 
@@ -41,17 +41,17 @@
     font-style: italic;
 } */
 
-.InfoTable.TemplateMetadata>div>div:nth-child(1) {
+.Main .InfoTable.TemplateMetadata>div>div:nth-child(1) {
     flex: 0 0 20em;
     text-align: left;
 }
 
-.InfoTable.TemplateMetadata>div>div:nth-child(2) {
+.Main .InfoTable.TemplateMetadata>div>div:nth-child(2) {
     flex: 0 0 10em;
     text-align: left;
 }
 
-.InfoTable.TemplateMetadata>div>div:nth-child(3) {
+.Main .InfoTable.TemplateMetadata>div>div:nth-child(3) {
     /* flex: 0 0 20em; */
     flex: 1 1 0px;
     text-align: left;
