+++ react-app/src/components/Metadata/view.tsx
@@ -5,228 +5,29 @@ import {
 
 import { Map as LeafletMap, Marker, Popup, TileLayer } from 'react-leaflet';
 
-import L from 'leaflet';
+import Leaflet from 'leaflet';
 import { PushpinFilled } from '@ant-design/icons';
 import ReactDOMServer from 'react-dom/server';
-import { MetadataField, Sample, Metadata } from '../sample/data';
+
+import { Sample, Metadata } from '../Main/data';
+import {
+    Template, GroupingLayout, FieldDefinitionsMap, FieldLayout
+} from '../../lib/comm/dynamicServices/SampleServiceClient';
 
 import './style.less';
 
-export interface FieldFormatBase {
-
-}
-
-export interface FloatFieldFormat extends FieldFormatBase {
-    precision?: number;
-    showThousandsSeparator?: boolean;
-}
-
-export interface IntegerFieldFormat extends FieldFormatBase {
-    showThousandsSeparator?: boolean;
-}
-
-export interface StringFieldFormat extends FieldFormatBase {
-}
-
-export interface DateFieldFormat extends FieldFormatBase {
-}
-
-export type FieldFormat = FloatFieldFormat | IntegerFieldFormat | StringFieldFormat | DateFieldFormat;
-
-export type FieldType = 'integer' | 'float' | 'string' | 'date';
-
-export interface LayoutFieldBase {
-    type: FieldType;
-    key: string;
-    description?: string;
-    units?: Array<string>;
-}
-
-export interface FloatField extends LayoutFieldBase {
-    type: 'float';
-    format?: FloatFieldFormat;
-}
-
-export interface IntegerField extends LayoutFieldBase {
-    type: 'integer';
-    format?: IntegerFieldFormat;
-}
-
-export interface StringField extends LayoutFieldBase {
-    type: 'string';
-    format?: StringFieldFormat;
-}
-
-export interface DateField extends LayoutFieldBase {
-    type: 'date';
-    format?: DateFieldFormat;
-}
-
-export type LayoutField = IntegerField | FloatField | StringField | DateField;
-
-export interface LayoutGroup {
-    key: string;
-    label: string;
-    description?: string;
-    fields: { [key in string]: LayoutField };
-    layout: Array<string>;
-
-}
-
-export type GroupLayout = Array<LayoutGroup>;
-
-export interface GroupSchema {
-    title: string;
-    layout: Array<string>;
-}
-
-export interface GroupsSchema {
-    description: GroupSchema;
-    collection: GroupSchema;
-    curation: GroupSchema;
-    geolocation: GroupSchema;
-}
-
-// TODO: 
-export type Unit = string;
-
-export interface FieldSchema {
-    key: string;
-    type: string;
-    group: string;
-    description?: string;
-    units?: Array<Unit>;
-    format?: any;
-}
-
-export interface FieldsSchema {
-    [k: string]: FieldSchema;
-}
-
-export interface Schema {
-    groups: GroupsSchema,
-    fields: FieldsSchema;
-}
-
-export interface WrappedMetadataValue {
-    type: string,
-    field: MetadataField;
-}
-
-const groupLayout: GroupLayout = [
-    {
-        key: 'description',
-        label: 'Description',
-        description: 'Fields which describe the overall sample event',
-        fields: {
-            'purpose': {
-                key: 'purpose',
-                type: 'string'
-            },
-            'material': {
-                key: 'material',
-                type: 'string'
-            }
-        },
-        layout: ['purpose', 'material']
-    },
-    {
-        key: 'collection',
-        label: 'Collection',
-        description: 'Fields which describe the collection',
-        fields: {
-            'collection_date': {
-                key: 'collection_date',
-                type: 'date',
-                description: 'Date upon which the sample was collected'
-            },
-            'collector_chief_scientist': {
-                key: 'collector_chief_scientist',
-                type: 'string'
-            },
-            'collection_method': {
-                key: 'collection_method',
-                type: 'string'
-            }
-        },
-        layout: ['collection_date', 'collector_chief_scientist', 'collection_method']
-    },
-    {
-        key: 'curation',
-        label: 'Curation',
-        description: 'Fields which describe the curation of the sample',
-        fields: {
-            'current_archive': {
-                key: 'current_archive',
-                type: 'string'
-            },
-            'current_archive_contact': {
-                key: 'current_archive_contact',
-                type: 'string'
-            }
-        },
-        layout: ['current_archive', 'current_archive_contact']
-    },
-    {
-        key: 'geolocation',
-        label: 'Geolocation',
-        description: 'Fields which describe the sample collection location',
-        fields: {
-            'coordinate_precision?': {
-                key: 'coordinate_precision?',
-                type: 'integer'
-            },
-            'latitude': {
-                key: 'latitude',
-                type: 'float',
-                units: ['degrees'],
-                format: {
-                    precision: 5
-                }
-            },
-            'longitude': {
-                key: 'longitude',
-                type: 'float',
-                format: {
-                    precision: 5
-                }
-            },
-            'navigation_type': {
-                key: 'navigation_type',
-                type: 'string'
-            },
-            'locality_description': {
-                key: 'locality_description',
-                type: 'string'
-            },
-            'location_description': {
-                key: 'location_description',
-                type: 'string'
-            },
-            'name_of_physiographic_feature': {
-                key: 'name_of_physiographic_feature',
-                type: 'string'
-            },
-            'primary_physiographic_feature': {
-                key: 'primary_physiographic_feature',
-                type: 'string'
-            }
-        },
-        layout: ['coordinate_precision?', 'latitude', 'longitude', 'navigation_type',
-            'locality_description', 'location_description', 'name_of_physiographic_feature',
-            'primary_physiographic_feature']
-    }
-];
-
 export interface SampleViewerProps {
     sample: Sample;
+    template: Template;
+    layout: GroupingLayout;
+    fields: FieldDefinitionsMap;
 }
 
 interface SampleViewerState {
 }
 
 export default class SampleViewer extends React.Component<SampleViewerProps, SampleViewerState> {
-    renderGeolocation(data: Metadata, group: LayoutGroup) {
+    renderGeolocation(data: Metadata) {
         const { latitude, longitude } = data;
         if (typeof latitude === 'undefined' || typeof longitude === 'undefined') {
             return <Alert type="warning" message="Both latitude and longitude must be present to display a map location" />;
@@ -234,11 +35,11 @@ export default class SampleViewer extends React.Component<SampleViewerProps, Sam
         const lat = latitude.value as number;
         const lng = longitude.value as number;
         const componentString = ReactDOMServer.renderToString(<PushpinFilled />);
-        const icon = L.divIcon({
+        const icon = Leaflet.divIcon({
             html: componentString,
             className: 'map-marker',
-            iconSize: L.point(20, 20),
-            tooltipAnchor: L.point(0, 10)
+            iconSize: Leaflet.point(20, 20),
+            tooltipAnchor: Leaflet.point(0, 10)
         });
         // const OpenTopoMap = L.tileLayer('https://{s}.tile.opentopomap.org/{z}/{x}/{y}.png', {
         //     maxZoom: 17,
@@ -268,10 +69,10 @@ export default class SampleViewer extends React.Component<SampleViewerProps, Sam
         </div>;
     }
 
-    renderControlledMetadataGroupExtras(data: Metadata, group: LayoutGroup) {
+    renderControlledMetadataGroupExtras(data: Metadata, group: FieldLayout) {
         switch (group.key) {
             case 'geolocation':
-                return this.renderGeolocation(data, group);
+                return this.renderGeolocation(data);
         }
     }
 
@@ -296,15 +97,13 @@ export default class SampleViewer extends React.Component<SampleViewerProps, Sam
         </div>;
     }
 
-
-
     renderControlledMetadataGrouped() {
         const sample = this.props.sample;
         const metadata = sample.metadata;
 
-        const rows = groupLayout.map((group) => {
+        const rows = this.props.layout.layout.map((group) => {
             const fields = group.layout.map((fieldName) => {
-                const field = group.fields[fieldName];
+                const field = this.props.fields[fieldName];
                 if (!field) {
                     console.warn('Field not found: ' + fieldName);
                     return null;
@@ -341,8 +140,6 @@ export default class SampleViewer extends React.Component<SampleViewerProps, Sam
                         {this.renderControlledMetadataGroupExtras(metadata, group)}
                     </div>
                 </div>
-
-
             </div>;
         });
 
@@ -351,20 +148,13 @@ export default class SampleViewer extends React.Component<SampleViewerProps, Sam
         </div>;
     }
 
-    renderNoCellData() {
-        return <div style={{
-            fontStyle: 'italic',
-            color: 'silver'
-        }}>-</div>;
-    }
-
     render() {
         const sample = this.props.sample;
         const metadata = sample.metadata;
 
-        const rows = groupLayout.map((group) => {
+        const rows = this.props.layout.layout.map((group) => {
             const fields = group.layout.map((fieldName) => {
-                const field = group.fields[fieldName];
+                const field = this.props.fields[fieldName];
                 if (!field) {
                     console.warn('Field not found: ' + fieldName);
                     return null;
@@ -391,7 +181,7 @@ export default class SampleViewer extends React.Component<SampleViewerProps, Sam
             } else {
                 content = <div style={{ fontStyle: 'italic' }}>No data</div>;
             }
-            return <div className="DataGroup">
+            return <div className="DataGroup" key={group.key}>
                 <div className="-title">
                     {group.label}
                 </div>
@@ -401,8 +191,6 @@ export default class SampleViewer extends React.Component<SampleViewerProps, Sam
                         {this.renderControlledMetadataGroupExtras(metadata, group)}
                     </div>
                 </div>
-
-
             </div>;
         });
