+++ react-app/src/components/NoData.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export class NoData extends React.Component<{}, {}> {
+    render() {
+        return <span style={{ color: 'gray' }}>-</span>;
+    }
+}
