+++ react-app/src/components/UserCard/view.test.data.json
@@ -1,0 +1,9 @@
+{
+    "users": [
+        {
+            "username": "kbaseuitest",
+            "realname": "KBase UI Test User",
+            "gravatarHash": "b4d95f8595104614355e6ee9c4c03e3f"
+        }
+    ]
+}
