+++ react-app/src/components/Versions.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ArrowRightOutlined } from '@ant-design/icons/lib/icons';
-import { Sample } from './sample/data';
+import { Sample } from './Main/data';
 import Select from 'antd/lib/select';
 import { Button } from 'antd';
 import UserCard from './UserCard/view';
