--- react-app/src/components/sample/style.less
@@ -1,111 +1,0 @@
-@import '../../base.less';
-
-.Sample {
-    flex: 1 1 0px;
-    display: flex;
-    flex-direction: column;
-    min-height: 0;
-}
-
-.Sample .ant-row {
-    flex: 1 1 0px;
-}
-
-.Nav-item {
-    border: 1px silver solid;
-    padding: 4px;
-}
-
-.Nav-item:hover {
-    background-color: silver;
-    cursor: pointer;
-}
-
-.InfoTable.UserMetadata>div>div:nth-child(1) {
-    flex: 0 0 20em;
-}
-
-.InfoTable.UserMetadata>div>div:nth-child(2) {
-    flex: 0 0 20em;
-    color: gray;
-    font-style: italic;
-}
-
-.InfoTable.ControlledMetadata>div>div:nth-child(1) {
-    flex: 0 0 20em;
-}
-
-/* .InfoTable.ControlledMetadata>div>div:nth-child(2) {
-    flex: 0 0 20em;
-    color: gray;
-    font-style: italic;
-} */
-
-.InfoTable.TemplateMetadata>div>div:nth-child(1) {
-    flex: 0 0 20em;
-    text-align: left;
-}
-
-.InfoTable.TemplateMetadata>div>div:nth-child(2) {
-    flex: 0 0 10em;
-    text-align: left;
-}
-
-.InfoTable.TemplateMetadata>div>div:nth-child(3) {
-    /* flex: 0 0 20em; */
-    flex: 1 1 0px;
-    text-align: left;
-}
-
-
-.map-marker {
-    color: red;
-    font-size: 20px;
-}
-
-.DataGroup {
-    display: flex;
-    flex-direction: row;
-}
-
-.DataGroup+.DataGroup {
-    margin-top: 20px;
-}
-
-.DataGroup>.-title {
-    font-weight: bold;
-    color: @gray-8;
-    font-size: 100%;
-    flex: 0 0 10em;
-}
-
-.DataGroup>.-body {
-    // margin-top: 10px;
-    flex: 1 1 0px;
-    display: flex;
-    flex-direction: column;
-}
-
-.DataGroup>.-body>.-extras {
-    margin-top: 10px;
-}
-
-.DataGroupSet {}
-
-.DataGroupSet+.DataGroupSet {
-    margin-top: 40px;
-}
-
-.DataGroupSet>.-title {
-    font-weight: bold;
-    color: @gray-8;
-    font-size: 110%;
-}
-
-.DataGroupSet>.-body {
-    margin-top: 10px;
-}
-
-.DataGroupSet>.-extras {
-    margin-top: 10px;
-}
