+++ react-app/src/data/sesar-template.json
@@ -1,0 +1,136 @@
+{
+    "id": "SESAR",
+    "header": [],
+    "columns": [
+        {
+            "type": "sample",
+            "key": "name",
+            "column": "Name"
+        },
+        {
+            "type": "sample",
+            "key": "sourceId",
+            "column": "IGSN"
+        },
+        {
+            "type": "sample",
+            "key": "sourceParentId",
+            "column": "Parent IGSN"
+        },
+        {
+            "type": "metadata",
+            "key": "release_date",
+            "column": "Release Date"
+        },
+        {
+            "type": "metadata",
+            "key": "material",
+            "column": "Material"
+        },
+        {
+            "type": "metadata",
+            "key": "field_name",
+            "column": "Field name (informal classification)"
+        },
+        {
+            "type": "metadata",
+            "key": "location_description",
+            "column": "Location Description"
+        },
+        {
+            "type": "metadata",
+            "key": "locality_description",
+            "column": "Locality Description"
+        },
+        {
+            "type": "metadata",
+            "key": "collection_method",
+            "column": "Collection method"
+        },
+        {
+            "type": "metadata",
+            "key": "purpose",
+            "column": "Purpose"
+        },
+        {
+            "type": "metadata",
+            "key": "latitude",
+            "column": "Latitude"
+        },
+        {
+            "type": "metadata",
+            "key": "longitude",
+            "column": "Longitude"
+        },
+        {
+            "type": "metadata",
+            "key": "coordinate_precision?",
+            "column": "Coordinate Precision"
+        },
+        {
+            "type": "metadata",
+            "key": "elevation_start",
+            "column": "Elevation start"
+        },
+        {
+            "type": "metadata",
+            "key": "elevation_unit",
+            "column": "Elevation unit"
+        },
+        {
+            "type": "metadata",
+            "key": "navigation_type",
+            "column": "Navigation type"
+        },
+        {
+            "type": "metadata",
+            "key": "primary_physiographic_feature",
+            "column": "Primary physiographic feature"
+        },
+        {
+            "type": "metadata",
+            "key": "name_of_physiographic_feature",
+            "column": "Name of physiographic feature"
+        },
+        {
+            "type": "metadata",
+            "key": "field_program_cruise",
+            "column": "Field program/cruise"
+        },
+        {
+            "type": "metadata",
+            "key": "collector_chief_scientist",
+            "column": "Collector/Chief Scientist"
+        },
+        {
+            "type": "metadata",
+            "key": "collection_date",
+            "column": "Collection date"
+        },
+        {
+            "type": "metadata",
+            "key": "collection_date_precision ",
+            "column": "Collection date precision"
+        },
+        {
+            "type": "metadata",
+            "key": "current_archive",
+            "column": "Current archive"
+        },
+        {
+            "type": "metadata",
+            "key": "current_archive_contact",
+            "column": "Current archive contact"
+        },
+        {
+            "type": "metadata",
+            "key": "related_identifiers",
+            "column": "Related Identifiers"
+        },
+        {
+            "type": "metadata",
+            "key": "relation_type",
+            "column": "Relation Type"
+        }
+    ]
+}
