+++ react-app/src/data/template-definitions.json
@@ -1,0 +1,10 @@
+{
+    "templates": [
+        {
+            "id": "SESAR",
+            "name": "System for Earth Sample Regsitration (SESAR)",
+            "description": "A sample template based on SESAR",
+            "reference": "https://www.geosamples.org/"
+        }
+    ]
+}
