+++ react-app/src/lib/comm/dynamicServices/data/grouping-layout.sesar.json
@@ -1,0 +1,50 @@
+{
+    "id": "sesear",
+    "name": "SESAR Layout",
+    "description": "default layout",
+    "layout": [
+        {
+            "key": "description",
+            "label": "Description",
+            "description": "Fields which describe the overall sample event",
+            "layout": [
+                "purpose",
+                "material"
+            ]
+        },
+        {
+            "key": "collection",
+            "label": "Collection",
+            "description": "Fields which describe the collection",
+            "layout": [
+                "collection_date",
+                "collector_chief_scientist",
+                "collection_method"
+            ]
+        },
+        {
+            "key": "curation",
+            "label": "Curation",
+            "description": "Fields which describe the curation of the sample",
+            "layout": [
+                "current_archive",
+                "current_archive_contact"
+            ]
+        },
+        {
+            "key": "geolocation",
+            "label": "Geolocation",
+            "description": "Fields which describe the sample collection location",
+            "layout": [
+                "coordinate_precision?",
+                "latitude",
+                "longitude",
+                "navigation_type",
+                "locality_description",
+                "location_description",
+                "name_of_physiographic_feature",
+                "primary_physiographic_feature"
+            ]
+        }
+    ]
+}
