+++ react-app/src/lib/comm/dynamicServices/data/metadata.definitions.json
@@ -1,0 +1,170 @@
+{
+    "fields": [
+        {
+            "key": "source_id",
+            "type": "string",
+            "kind": "registration",
+            "label": "Source Id"
+        },
+        {
+            "key": "source_parent_id",
+            "type": "string",
+            "kind": "registration",
+            "label": "Parent source id"
+        },
+        {
+            "key": "id",
+            "type": "string",
+            "kind": "registration",
+            "label": "ID"
+        },
+        {
+            "key": "name",
+            "type": "string",
+            "kind": "registration",
+            "label": "Name"
+        },
+        {
+            "key": "purpose",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Purpose"
+        },
+        {
+            "key": "material",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Material"
+        },
+        {
+            "key": "collection_date",
+            "type": "date",
+            "kind": "descriptive",
+            "description": "Date upon which the sample was collected",
+            "label": "Collection Date"
+        },
+        {
+            "key": "collection_date_precision",
+            "type": "integer",
+            "kind": "descriptive",
+            "label": "Collection date precision"
+        },
+        {
+            "key": "collector_chief_scientist",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Collector Chief Scientist"
+        },
+        {
+            "key": "collection_method",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Collection Methohd"
+        },
+        {
+            "key": "release_date",
+            "type": "date",
+            "kind": "descriptive",
+            "label": "Release Date"
+        },
+        {
+            "key": "field_name",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "field Name"
+        },
+        {
+            "key": "elevation_start",
+            "type": "float",
+            "kind": "descriptive",
+            "label": "Elevation start"
+        },
+        {
+            "key": "elevation_unit",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Elevation unit"
+        },
+        {
+            "key": "field_program_cruise",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Field program / cruise"
+        },
+        {
+            "key": "current_archive",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Current Archive"
+        },
+        {
+            "key": "current_archive_contact",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Current Archive Contact"
+        },
+        {
+            "key": "coordinate_precision?",
+            "type": "integer",
+            "kind": "descriptive",
+            "label": "Coordinate Precision"
+        },
+        {
+            "key": "latitude",
+            "type": "float",
+            "kind": "descriptive",
+            "label": "Latitude",
+            "units": "degrees",
+            "precision": 5
+        },
+        {
+            "key": "longitude",
+            "type": "float",
+            "kind": "descriptive",
+            "label": "Longitude",
+            "precision": 5
+        },
+        {
+            "key": "navigation_type",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Navigation Type"
+        },
+        {
+            "key": "locality_description",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Locality Description"
+        },
+        {
+            "key": "location_description",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Location Description"
+        },
+        {
+            "key": "name_of_physiographic_feature",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Name of Physiographic Feature"
+        },
+        {
+            "key": "primary_physiographic_feature",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Primary Physiographic Feature"
+        },
+        {
+            "key": "related_identifiers",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Related Identifiers"
+        },
+        {
+            "key": "relation_type",
+            "type": "string",
+            "kind": "descriptive",
+            "label": "Relation Type"
+        }
+    ]
+}
