+++ react-app/src/lib/comm/dynamicServices/data/sesar-template.json
@@ -1,0 +1,35 @@
+{
+    "id": "SESAR",
+    "version": 1,
+    "created_at": 0,
+    "created_by": "eapearson",
+    "header": [],
+    "columns": [
+        "name",
+        "source_id",
+        "source_parent_id",
+        "release_date",
+        "material",
+        "field_name",
+        "location_description",
+        "locality_description",
+        "collection_method",
+        "purpose",
+        "latitude",
+        "longitude",
+        "coordinate_precision?",
+        "elevation_start",
+        "elevation_unit",
+        "navigation_type",
+        "primary_physiographic_feature",
+        "name_of_physiographic_feature",
+        "field_program_cruise",
+        "collector_chief_scientist",
+        "collection_date",
+        "collection_date_precision",
+        "current_archive",
+        "current_archive_contact",
+        "related_identifiers",
+        "relation_type"
+    ]
+}
